+++ internal/provider/resource_iosxe_cts_test.go
@@ -60,6 +60,7 @@ func TestAccIosxeCTS(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_cts.test", "sxp_listener_hold_max_time", "300"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_cts.test", "role_based_enforcement", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_cts.test", "role_based_enforcement_logging_interval", "300"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_cts.test", "credentials_id", "DEVICE123"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -76,7 +77,7 @@ func TestAccIosxeCTS(t *testing.T) {
 				ImportState:             true,
 				ImportStateVerify:       true,
 				ImportStateIdFunc:       iosxeCTSImportStateIdFunc("iosxe_cts.test"),
-				ImportStateVerifyIgnore: []string{},
+				ImportStateVerifyIgnore: []string{"credentials_password_set"},
 				Check:                   resource.ComposeTestCheckFunc(checks...),
 			},
 		},
@@ -157,6 +158,9 @@ func testAccIosxeCTSConfig_all() string {
 	config += `	sxp_listener_hold_max_time = 300` + "\n"
 	config += `	role_based_enforcement = true` + "\n"
 	config += `	role_based_enforcement_logging_interval = 300` + "\n"
+	config += `	credentials_id = "DEVICE123"` + "\n"
+	config += `	credentials_password_type = "0"` + "\n"
+	config += `	credentials_password = "MySecretPassword123"` + "\n"
 	config += `	depends_on = [iosxe_restconf.PreReq0, ]` + "\n"
 	config += `}` + "\n"
 	return config