+++ docs/guides/changelog.md
@@ -9,6 +9,7 @@ description: |-
 
 ## 0.10.0 (unreleased)
 
+- Add `credentials_id`, `credentials_password`, `credentials_password_type` attributes to `iosxe_cts` resource and data source for TrustSec NAD credentials
 - Add `flooding_suppression_address_resolution_disable` attribute to `iosxe_evpn` resource and data source
 - Add `ip_domain_lookup_nsap`, `ip_domain_lookup_recursive`, and `ip_domain_lookup_vrfs*` attributes to `iosxe_system` resource and data source
 - Add `iosxe_evpn_ethernet_segment` resource and data source for managing L2VPN EVPN Ethernet Segment configuration